+++ ml-agents-envs/mlagents_envs/envs/unity_pettingzoo_base_env.py
@@ -132,7 +132,7 @@ class UnityPettingzooBaseEnv:
                         continue
                 if act_spec.continuous_size > 0:
                     c_space = spaces.Box(
-                        -1, 1, (act_spec.continuous_size,), dtype=np.int32
+                        -1, 1, (act_spec.continuous_size,), dtype=np.float32
                     )
                     if self._seed is not None:
                         c_space.seed(self._seed)
